Ext4Pkg: Replace EXT4_BLOCK_NR with 32-bit EXT2_BLOCK_NR in BlockMap
Drop unnecessary casts, because we consider BlockMap is used for 32-bit
fs ext2/3. Also we consider 32-bit LogicalBlock for BlockMap
functionality, so make a cast in it's caller Ext4GetExtent

+++ Ext4Pkg/Ext4Dxe/Extents.c
@@ -259,7 +259,8 @@ Ext4GetExtent (
 
   if (!(Inode->i_flags & EXT4_EXTENTS_FL)) {
     // If this is an older ext2/ext3 filesystem, emulate Ext4GetExtent using the block map
-    Status = Ext4GetBlocks (Partition, File, LogicalBlock, Extent);
+    // We cast LogicalBlock to UINT32, considering ext2/3 are 32-bit
+    Status = Ext4GetBlocks (Partition, File, (UINT32) LogicalBlock, Extent);
 
     if (!EFI_ERROR (Status)) {
       Ext4CacheExtents (File, Extent, 1);
